code/requirements.txt
geopandas==0.9.0
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
Shapely==1.8.5.post1
